scraper/requirements.txt
BeautifulSoup4
requests
python-dotenv
boto3
pandas